backend/requirements.txt
# TelePlay Backend

# Telegram MTProto Client
pyrotgfork>=2.2.0
TgCrypto>=1.2.3

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0

# Rate Limiting
slowapi>=0.1.9


